Update PropertiesPage.xaml
Please note that as I was unable to run the application, it was very difficult for to properly determine how much space between padding and margins would be required for each section. (Postcode, Address, Price, Picture).

However the basic principle for this objective from my view is to extend the Child(Property) between the padding and the margins to allow for more space, then simply put it the additional information and draw it using Bindings from the SearchResult.cs. I did understand the principle of this objective and how it was to be carried out. However since I was unable to load the application due to emulator problems and general problems with visual studio, it was difficult for me to accurately understand how much spacing would be needed, as I wrote this code blind and was unable to load the application to see the results. Tweaking may be required to certain numbers in regards to the padding and layout.